tests/wpt/tests/css/css-masking/parsing/mask-repeat-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: parsing mask-repeat with invalid values</title>
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-mask-repeat">
<meta name="assert" content="mask-repeat supports only the grammar.">
<!--
<mask-repeat> = <repeat-style>#
<repeat-style> = repeat-x | repeat-y | [repeat | space | round | no-repeat]{1,2}
 -->
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("mask-repeat", "auto");
test_invalid_value("mask-repeat", "repeat-z");
test_invalid_value("mask-repeat", "repeat undefined");
test_invalid_value("mask-repeat", "undefined repeat");
test_invalid_value("mask-repeat", "repeat; round");
</script>
</body>
</html>
